Build fails on OpenVMS
I've tried to build Text::CSV_XS in the following environment:

Text::CSV_XS 0.80
Perl 5.12.2
OpenVMS 8.3-1H1

Running Makefile.PL works fine, but the generated makefile (descrip.mms
on OpenVMS) can not be parsed successfully and therefore the module does
not build.

This is caused by the following line

cover test_cover:

Multiple targets on one line have to be separated by a , in OpenVMS.
This is not a problem with MakeMaker, but gets inserted directly by
Makefile.PL

The attached patch fixes the problem (conditionally adds a coma in case
the script runs on OpenVMS, also adds a space before the colon needed in
list mode). This should not have any negative impact on other platforms.
It has been successfully cross tested on Linux (openSUSE 11.3).

I'd be happy if your would consider this change for inclusion in the
next version.

Greetings, Martin

Signed-off-by: H.Merijn Brand <merijn@tux09.procura.nl>