Testing *my* module should not break on failing other modules
Encode is only used for the *test*, not required for production,
so if Encode is too old to cope, don't run the test.

Maybe 2.47 is not the lowest required version, but it installed
fin on all my perl versions from 5.8.0 and up. As Unicode is only
tested on 5.8.2 and up, I'm safe in this recommendation