avoid producing double references on per-line AVs in xsParse_all ()
The code used newRV (), which is an alias for newRV_inc (), hence the
line AVs each had a reference count of 2 when they were returned.

rav_free () handled this by releasing both the RV and the underlying
AV, but this is unnecessary when the AVs have the correct reference
count.