Cache not re-read on getline_all
When detecting \r as eol, the cache is updated, but not re-used when
getline_all is used, assuming the struct itself was already updated.

It might be, but it is not guaranteed. With \r, eol is set to \r, but
some flags, including eil_is_cr is set on re-evaluation of the cache