Quote strings:
Standard Perl does not accept the akward auto-quoting rules for strings.
If it's a string, it needs to be a string.

Rules for quoto-quoting include stuff like:

* If there's a new line before the fat comma, it's not a string
* If it starts with a lowercase p, it might change
* If there is a plus, it's not a string
* If there's a minu, it's still a string

In short, this entire thing is confusing and problematic and even tidying
can break it.

This commit simply quotes all strings, either used in a hash definition or
in a hash key.